changed place of comments
