+++ app/assets/stylesheets/components/all.scss
@@ -5,3 +5,4 @@
 @import "user";
 @import "article";
 @import "company";
+@import "comment";
