+++ app/assets/stylesheets/components/comment.scss
@@ -1,0 +1,3 @@
+.comment-photo {
+  width: 100%;
+}
