+++ app/decorators/article_decorator.rb
@@ -1,7 +1,7 @@
 class ArticleDecorator < ApplicationDecorator
   include ActionView::Helpers::DateHelper
 
-  delegate :id, :content, :title, :status
+  delegate :id, :content, :title, :status, :comments
 
   def publication
     "Published #{distance_of_time_in_words_to_now(object.updated_at)} ago"
