+++ app/views/comments/_form.html.slim
@@ -1,4 +1,4 @@
-= simple_form_for [:admin_scope, article, comment] do |f|
+= simple_form_for [article, comment] do |f|
   .form-inputs
     = f.input :content
